move TagNames into ServiceRemotingTags
